Clean up usage of Steam App IDs, and implement Game ID. Steam's Game ID is a 64-bit representation of an app, usually only used locally. The lower 32- bits are just the App ID, but the rest can be used to identify a mod, or a non-Steam shortcut. In the case of a mod, the App ID portion is still used for Steam to know which app to launch. In most places, a Game ID can be used in place of an App ID since they have the same value for "normal" apps.
This shouldn't have any net change, but will allow implementation of Source
mod support in the Playnite Steam plugin.

* Fix usages of App IDs to be uint instead of int, to match Steam.
* Implement Steam Game ID for local usages (not online).
* Switch from steam://run to rungame, to support running by Game ID.
* Switch Playnite Game ID to use Steam Game ID instead of App ID for more
clear identification of app.

+++ source/Plugins/SteamLibrary.Tests/SteamLibraryTests.cs
@@ -68,7 +68,7 @@ namespace SteamLibrary.Tests
         {
             var steamLib = CreateLibrary();
             var games = steamLib.GetInstalledGames();
-            var state = Steam.GetAppState(int.Parse(games.Values.First().GameId));
+            var state = Steam.GetAppState(ulong.Parse(games.Values.First().GameId));
             Assert.IsTrue(state.Installed);
         }
     }
